% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begg.test.R
\name{begg.test}
\alias{begg.test}
\title{Begg's test of publication bias, standard form or bias-corrected}
\usage{
begg.test(y, v, adjusted = TRUE, ...)
}
\arguments{
\item{y}{study effects}

\item{v}{study effect variances}

\item{corrected}{whether to adjust the asymptotic variance to an estimate of the correct value (defaults to TRUE)}
}
\value{
a named list of class "htest"
}
\description{
Begg's test of publication bias, standard form or bias-corrected
}
\examples{

n <- 10
y <- rnorm(n)
v <- runif(n)
begg.test(y,v)
begg.test(y,v,adjusted=FALSE)
}
